Blockchain Identity And Voting Systems
### Blockchain Identity And Voting Systems Can Ensure Fair Elections
By allowing voters to see the number of votes that are received, and making it transparent and trustworthy as an electoral system, everyone will be able to "recognize" the legitimacy of the vote. In postal voting, verifying the identity of voters and states that guarantee the traceability of each vote is crucial to preventing voter fraud.